backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
# django-colorfield==0.11.0
django-extra-fields==3.0.2
psycopg2-binary==2.9.9